+++ crates/metadata/src/map/mob.rs
@@ -76,7 +76,7 @@ pub fn get_mob_wz_info(mob_life: serde_json::Value) -> Result<BaseMob, MapMetada
     let fs: f32 = info["fs"].as_i64().unwrap_or(0) as f32;
     let level: i16 = info["level"].as_i64().unwrap_or(0) as i16;
     let max_hp: i32 = info["maxHP"].as_i64().unwrap_or(0) as i32;
-    let max_mp: i32 = info["maxMP"].as_i64().unwrap_or(0) as i32;
+    let max_mp: i16 = info["maxMP"].as_i64().unwrap_or(0) as i16;
     let mob_type: i16 = info["mobType"].as_i64().unwrap_or(0) as i16;
     let pushed: i8 = info["pushed"].as_i64().unwrap_or(0) as i8;
     let speed: i16 = info["speed"].as_i64().unwrap_or(0) as i16;
